Big and necessary sacalability issues resolved:     - Completely changed the way Elements are handled: Now there's a base class called StoryElement, which doesn't hold pointers to documents, a sub-class calles TangibleElement, which holds the pointers. There are no pre-defined wxString attributes in Characters, Locations and Items anymore, which means all attributes are treated equally, be it custom or not (using std::multimap<wxString, wxString> or its typdef, wxStringMultimap).     - ElementShowcases have followed up with similar changes and there are no more pre-defined wxString attributes. short attributes use a wxWrapSizer so that they wrap in lines.     - The database now only has the 'story_elements' table, which specifies the name of the Element and the C++ class (RTTI is used when loading them). Everything else is serialized in XML using xsSerializable.     - Started using polymorphism for EVERYTHING involving Elements (I think), so there's now only a DoEditStoryElement() function in amProjectManager instead of one for each Element. This also applies to creating Elements, etc. Also outsourced Editing to amProjectManager::StartEditingStoryElement().

+++ src/Overview.cpp
@@ -41,7 +41,7 @@ amBookPicker::amBookPicker(amOverview* parent) :
 	SetSizerAndFit(vertical);
 }
 
-void amBookPicker::AddButton(Book* book, int index)
+void amBookPicker::AddButton(Book* book, size_t index)
 {
 	amBookButton* button = new amBookButton(m_mainWindow, book,
 		amGetScaledImage(m_bookCoverSize.x, m_bookCoverSize.y, book->cover),
@@ -81,7 +81,7 @@ void amBookPicker::OnBookClicked(wxCommandEvent& event)
 
 void amBookPicker::SetSelectionByBook(Book* book)
 {
-	for ( int i = 0; i < m_bookButtons.size(); i++ )
+	for ( size_t i = 0; i < m_bookButtons.size(); i++ )
 	{
 		if ( book == m_bookButtons[i]->book )
 			m_bookButtons[i]->SetValue(true);
